requirements.txt
smbus2
pyserial
